init PrerequisiteList
