+++ src/app/(admin)/admin/programs/prerequisite/PrerequisiteDetails.tsx
@@ -18,6 +18,7 @@ import { db } from '@/lib/config/firebase';
 import { doc, getDoc, onSnapshot } from 'firebase/firestore';
 import PrerequisiteForm from './PrerequisiteForm';
 import { Prerequisite, Program } from '../modal/program';
+import PrerequisiteList from './PrerequisiteList';
 
 export default function PrerequisiteDetails(props: BoxProps) {
   const [certificateId] = useQueryState('certificate');
@@ -41,30 +42,14 @@ export default function PrerequisiteDetails(props: BoxProps) {
       {loading ? (
         <Loader />
       ) : (
-        certificate && <PrerequisiteView certificate={certificate} />
+        certificate && (
+          <>
+            <PrerequisiteForm certificate={certificate} />
+            <PrerequisiteList />
+          </>
+        )
       )}
     </Box>
-  );
-}
-
-function PrerequisiteView({ certificate }: { certificate: Certificate }) {
-  const [data, setData] = useState<Prerequisite[]>();
-  const [programId] = useQueryState('id');
-  const [program, setProgram] = useState<Program>();
-
-  useEffect(() => {
-    if (programId) {
-      const unsubscribe = onSnapshot(doc(db, 'programs', programId), (doc) => {
-        setProgram({ ...doc.data(), id: doc.id } as Program);
-      });
-      return () => unsubscribe();
-    }
-  }, [programId]);
-
-  return (
-    <>
-      <PrerequisiteForm certificate={certificate} />
-    </>
   );
 }
